% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_schools_utils.R
\name{assign_vaccination_ids}
\alias{assign_vaccination_ids}
\title{Assign unique IDs to vaccination data rows}
\usage{
assign_vaccination_ids(vax_data)
}
\arguments{
\item{vax_data}{A data frame of school-level vaccination data.}
}
\value{
The same data frame with a new column `vacc_school_id`.
}
\description{
Assign unique IDs to vaccination data rows
}
