% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{combine_vaxview_data}
\alias{combine_vaxview_data}
\title{Combine child, school, and teen vaccine coverage data}
\usage{
combine_vaxview_data(child_dt, school_dt, teen_dt)
}
\arguments{
\item{child_dt}{Child-level data.table}

\item{school_dt}{School-level data.table}

\item{teen_dt}{Teen-level data.table}
}
\value{
A unified data.table of vaccine coverage
}
\description{
Combine child, school, and teen vaccine coverage data
}
