% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greatschools_utils.R
\name{extract_gs_schoolsdata}
\alias{extract_gs_schoolsdata}
\title{Extract schools data from GreatSchools.org for a single zip code}
\usage{
extract_gs_schoolsdata(zipcode = 21231)
}
\arguments{
\item{zipcode}{A US zip code (default = 21231)}
}
\value{
A data.frame of schools data from GreatSchools.org
}
\description{
Extract schools data from GreatSchools.org for a single zip code
}
\details{
This function fetches the schools data for a given zip code from GreatSchools.org
by making an HTTP GET request to the search page and extracting the embedded JSON data.
The schools data is returned as a data.frame.
}
