% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greatschools_utils.R
\name{get_county_zips}
\alias{get_county_zips}
\title{Get county-zip mapping with FIPS codes}
\usage{
get_county_zips()
}
\value{
A data.frame mapping zip codes to counties with FIPS codes
}
\description{
Get a mapping of zip codes to counties with FIPS codes
}
\examples{
county_zips <- get_county_zips()

}
