% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_sd_from_ci}
\alias{get_sd_from_ci}
\title{Get SD from Confidence Interval}
\usage{
get_sd_from_ci(lower_ci, upper_ci, n, conf_level = 0.95)
}
\arguments{
\item{lower_ci}{Lower confidence interval}

\item{upper_ci}{Upper confidence interval}

\item{n}{Sample size}

\item{conf_level}{Confidence level (default 0.95)}
}
\value{
Standard deviation
}
\description{
Get SD from Confidence Interval
}
\examples{
get_sd_from_ci(40, 60, 100)
get_sd_from_ci(30, 70, 50, conf_level = 0.99)

}
