% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding_utils.R
\name{get_zip}
\alias{get_zip}
\title{Get ZIP Code (ZCTA) from Latitude and Longitude}
\usage{
get_zip(lat, lon, data_year = 2020)
}
\arguments{
\item{lat}{Numeric vector of latitudes in decimal degrees (WGS84).}

\item{lon}{Numeric vector of longitudes in decimal degrees (WGS84).}
}
\value{
A tibble containing:
\describe{
  \item{latitude}{Input latitude}
  \item{longitude}{Input longitude}
  \item{zip_code}{5-digit ZCTA code}
}
}
\description{
Uses Census ZCTA polygons to identify the ZIP Code Tabulation Area (ZCTA)
containing each latitude/longitude point.
}
\examples{
get_zip(39.2904, -76.6122)

}
