% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{load_school_vaxview}
\alias{load_school_vaxview}
\title{Load and process school-level vaccine coverage data}
\usage{
load_school_vaxview(
  path = "model/data/nis/SchoolVaxView.csv",
  state_name,
  pull_new = TRUE
)
}
\arguments{
\item{path}{Path to the SchoolVaxView CSV file}

\item{state_name}{Full state name}
}
\value{
A data.table with processed school-level vaccine data for all available vaccines
}
\description{
Load and process school-level vaccine coverage data
}
