% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{load_teen_vaxview}
\alias{load_teen_vaxview}
\title{Load and process teen-level vaccine coverage data}
\usage{
load_teen_vaxview(
  path = "model/data/nis/TeenVaxView.csv",
  state_name,
  pull_new = TRUE
)
}
\arguments{
\item{path}{Path to the TeenVaxView CSV file}

\item{state_name}{Full state name}
}
\value{
A data.table with processed teen-level vaccine data for all available vaccines
}
\description{
Load and process teen-level vaccine coverage data
}
