% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_schools_utils.R
\name{match_locs_level2}
\alias{match_locs_level2}
\alias{match_locations}
\title{match_locs_level2}
\usage{
match_locations(
  a,
  names,
  return_name = TRUE,
  return_score = FALSE,
  return_score_matrix = FALSE
)
}
\arguments{
\item{a}{location name to match}

\item{names}{location names to match against}

\item{return_name}{TRUE/FALSE return standardized name}

\item{return_score}{TRUE/FALSE}

\item{return_score_matrix}{TRUE/FALSE}

\item{return_Code}{TRUE/FALSE}
}
\value{
ISOs, country names, matching scores, full matching distance matrix
}
\description{
use stringdist to get best match for country name, if not official country
}
