% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greatschools_utils.R
\name{process_zip}
\alias{process_zip}
\title{Function to process one ZIP}
\usage{
process_zip(zipcode)
}
\arguments{
\item{zipcode}{}
}
\value{
A data.frame of schools data for the given ZIP code, or NULL if no schools found
}
\description{
Function to process one ZIP
}
