% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{save_vaxview_parquet}
\alias{save_vaxview_parquet}
\title{Save vaccine coverage data to Parquet format}
\usage{
save_vaxview_parquet(vv_all, path)
}
\arguments{
\item{vv_all}{Combined vaccine coverage data.table}

\item{path}{Path to save the Parquet file}
}
\description{
Save vaccine coverage data to Parquet format
}
