% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{setup_vaxview_data}
\alias{setup_vaxview_data}
\title{set up vaxview data}
\usage{
setup_vaxview_data(
  state_name,
  child_vaxview_path,
  school_vaxview_path,
  teen_vaxview_path,
  vax_data_path,
  pull_new = TRUE
)
}
\arguments{
\item{state_name}{Full state name}

\item{child_vaxview_path}{Path to child VaxView CSV}

\item{school_vaxview_path}{Path to school VaxView CSV}

\item{teen_vaxview_path}{Path to teen VaxView CSV}

\item{vax_data_path}{Path to save combined vaccine data Parquet}
}
\value{
A data.table of combined vaccine coverage data
}
\description{
set up vaxview data
}
