% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_schools_utils.R
\name{standardize_location_strings}
\alias{standardize_location_strings}
\title{standardize_location_strings}
\usage{
standardize_location_strings(location_name)
}
\arguments{
\item{location_name}{location name to match}
}
\value{
standardized string of the location name.
}
\description{
Standardize each level of a string name. This is used by others functions that standardize and match names.
}
