% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_schools_utils.R
\name{standardize_string}
\alias{standardize_string}
\title{standardize_string}
\usage{
standardize_string(string)
}
\arguments{
\item{string}{string location name to standardize.}
}
\value{
standardized string of the location names.
}
\description{
Standardize indivual levels of a location of grouped names, separated by "|", for example "DC|Maryland".
This is used by standardize_location_strings.
}
