% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyschoolvax-package.R
\docType{package}
\name{tidyschoolvax-package}
\alias{tidyschoolvax}
\alias{tidyschoolvax-package}
\title{tidyschoolvax: Preprocessing and Data Quality Assurance for School Vaccination Data}
\description{
Provides modular functions for preprocessing, cleaning, and quality assurance of state-level school vaccination data. The package standardizes data formats across states and implements comprehensive data quality checks for measles, mumps, and rubella (MMR) vaccination coverage data from kindergarten enrollment records.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ACCIDDA/tidyschoolvax}
  \item Report bugs at \url{https://github.com/ACCIDDA/tidyschoolvax/issues}
}

}
\author{
\strong{Maintainer}: ACCIDDA Team \email{contact@accidda.org}

}
